+++ pkg/deployment/member/phase_updates.go
@@ -104,6 +104,7 @@ func removeMemberConditionsMapFunc(m *api.MemberStatus) {
 	m.Conditions.Remove(api.ConditionTypePVCResizePending)
 	m.Conditions.Remove(api.ConditionTypeArchitectureMismatch)
 	m.Conditions.Remove(api.ConditionTypeArchitectureChangeCannotBeApplied)
+	m.Conditions.Remove(api.ConditionTypeMemberVolumeUnschedulable)
 
 	m.RemoveTerminationsBefore(time.Now().Add(-1 * recentTerminationsKeepPeriod))